+++ lona_picocss/templates/picocss/header.html
@@ -1,19 +1,19 @@
-{% set resolve_url = Lona.import('lona_picocss.themes.resolve_url') %}
+{% set resolve_url = Lona.import('lona_picocss.settings.resolve_url') %}
 <nav>
     <ul>
-        <li><a href="/" id="brand">{% if theme_data.theme_settings.logo %}<img src="{{ Lona.load_static_file(theme_data.theme_settings.logo) }}"> {% endif %}{% if theme_data.theme_settings.brand %}<strong>{{ theme_data.theme_settings.brand }}</strong>{% endif %}</a></li>
+        <li><a href="/" id="brand">{% if theme_data.logo_path %}<img src="{{ theme_data.logo_path }}"> {% endif %}{% if theme_data.theme_settings.brand %}<strong>{{ theme_data.theme_settings.brand }}</strong>{% endif %}</a></li>
     </ul>
     <ul>
         {% for label, target in theme_data.menu %}
             {% if isinstance(target, str) %}
-                <li><a href="{{ resolve_url(Lona, target) }}">{{ label }}</a></li>
+                <li><a href="{{ resolve_url(target) }}">{{ label }}</a></li>
             {% elif isinstance(target, list) %}
                 <li>
                     <details role="list" dir="rtl">
                         <summary aria-haspopup="listbox" role="link">{{ label }}</summary>
                         <ul role="listbox">
                             {% for sub_label, sub_target in target %}
-                                <li><a href="{{ resolve_url(Lona, sub_target) }}">{{ sub_label }}</a></li>
+                                <li><a href="{{ resolve_url(sub_target) }}">{{ sub_label }}</a></li>
                             {% endfor %}
                         </ul>
                     </details>
