+++ lona_picocss/views/settings.py
@@ -38,18 +38,21 @@ class SettingsView(View):
 
     def handle_change(self, input_event):
         name = input_event.node.state['name']
+        value = input_event.node.value
 
-        setattr(self.server.settings, name, input_event.node.value)
+        settings.set(name, value)
+        settings.render_theme()
 
         return self.refresh()
 
-
     def apply(self, input_event):
+        settings.render_theme()
+
         return self.refresh()
 
     def reset(self, input_event):
-        for key, value in settings.DEFAULTS.items():
-            setattr(self.server.settings, key, value)
+        settings.reset()
+        settings.render_theme()
 
         return self.refresh()
 
@@ -110,15 +113,14 @@ class SettingsView(View):
         settings_pre = html.query_selector('pre')
         settings_form = html[2][0]
 
-        for index, name in enumerate(settings.SETTINGS.keys()):
+        for index, name in enumerate(settings.get_names()):
             slug = name[len('PICOCSS_'):].replace('_', '-').lower()
             verbose_name = name[len('PICOCSS_'):].replace('_', ' ').title()
-            default = settings.DEFAULTS[name]
-            values = settings.SETTINGS[name]
-            value = self.server.settings.get(name, default)
+            value = settings.get(name)
+            values = settings.get_default(name, value)
 
-            # object values
-            if name in settings.OBJECT_SETTINGS:
+            # unsupported settings
+            if name in ('PICOCSS_MENU', ):
                 continue
 
             # string values
@@ -176,7 +178,12 @@ class SettingsView(View):
                 settings_form.append(Br())
 
             # add config variable to compiled output
-            if value != settings.DEFAULTS[name]:
+            default = settings.get_default(name)
+
+            if isinstance(default, list) and len(default) > 0:
+                default = default[0]
+
+            if value != default:
                 settings_pre.write_line(
                     f'{settings_prefix}{name} = {repr(value)}',
                 )
